+++ CHANGELOG.rst
@@ -15,6 +15,14 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.1
+---
+2019-09-08
+
+- Documentation fixes.
+- Speed up tests.
+- Clean up requirements.
+
 0.0.13
 ------
 2019-09-07